+++ backend/canisters/community/api/src/updates/mod.rs
@@ -1,4 +1,3 @@
-pub mod accept_rules;
 pub mod add_members_to_channel;
 pub mod add_reaction;
 pub mod block_user;