+++ backend/canisters/community/impl/src/model/channels.rs
@@ -152,6 +152,7 @@ impl Channel {
                 GroupPermissions::default(),
                 None,
                 None,
+                false,
                 now,
             ),
             date_imported: None,
@@ -260,28 +261,23 @@ impl Channel {
         let can_view_latest_message = self.can_view_latest_message(member.is_some(), is_community_member, is_public_community);
         let updates_from_events = chat.summary_updates_from_events(since, user_id, now);
 
-        let membership = member.map(|m| {
-            let mut rules_accepted = None;
-            if let Some(accepted) = &m.rules_accepted {
-                if updates_from_events.rules_changed || accepted.timestamp > since {
-                    rules_accepted = Some(accepted.value >= chat.rules.text.version);
-                }
-            }
-
-            ChannelMembershipUpdates {
-                role: updates_from_events.role_changed.then_some(m.role.into()),
-                mentions: updates_from_events.mentions,
-                notifications_muted: m.notifications_muted.if_set_after(since).cloned(),
-                my_metrics: self.chat.events.user_metrics(&m.user_id, Some(since)).map(|m| m.hydrate()),
-                latest_threads: self.chat.events.latest_threads(
-                    m.min_visible_event_index(),
-                    m.threads.iter(),
-                    Some(since),
-                    MAX_THREADS_IN_SUMMARY,
-                    now,
-                ),
-                rules_accepted,
-            }
+        let membership = member.map(|m| ChannelMembershipUpdates {
+            role: updates_from_events.role_changed.then_some(m.role.into()),
+            mentions: updates_from_events.mentions,
+            notifications_muted: m.notifications_muted.if_set_after(since).cloned(),
+            my_metrics: self.chat.events.user_metrics(&m.user_id, Some(since)).map(|m| m.hydrate()),
+            latest_threads: self.chat.events.latest_threads(
+                m.min_visible_event_index(),
+                m.threads.iter(),
+                Some(since),
+                MAX_THREADS_IN_SUMMARY,
+                now,
+            ),
+            rules_accepted: m
+                .rules_accepted
+                .as_ref()
+                .filter(|accepted| updates_from_events.rules_changed || accepted.timestamp > since)
+                .map(|accepted| accepted.value >= chat.rules.text.version),
         });
 
         ChannelUpdates::Updated(CommunityCanisterChannelSummaryUpdates {