+++ backend/canisters/community/impl/src/queries/selected_initial.rs
@@ -24,5 +24,6 @@ fn selected_initial_impl(args: Args, state: &RuntimeState) -> Response {
         blocked_users: members.blocked(),
         invited_users: state.data.invited_users.users(),
         rules: state.data.rules.clone().into(),
+        access_rules: state.data.rules.clone().into(),
     })
 }