+++ backend/canisters/user/api/src/updates/send_message_with_transfer_to_group.rs
@@ -2,7 +2,7 @@ use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{
     ChatId, CompletedCryptoTransaction, Cryptocurrency, EventIndex, GroupReplyContext, MessageContentInitial, MessageId,
-    MessageIndex, TimestampMillis, User,
+    MessageIndex, TimestampMillis, User, Version,
 };
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -15,6 +15,8 @@ pub struct Args {
     pub replies_to: Option<GroupReplyContext>,
     pub mentioned: Vec<User>,
     pub correlation_id: u64,
+    #[serde(default)]
+    pub rules_accepted: Option<Version>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -29,6 +31,7 @@ pub enum Response {
     TransferCannotBeZero,
     UserSuspended,
     ChatFrozen,
+    RulesNotAccepted,
     InternalError(String, CompletedCryptoTransaction),
 }