+++ backend/canisters/user_index/impl/src/timer_job_types.rs
@@ -204,6 +204,7 @@ impl Job for JoinUserToGroup {
                     invite_code: None,
                     correlation_id: 0,
                     is_platform_moderator: state.data.platform_moderators.contains(&self.user_id),
+                    is_bot: u.is_bot,
                 })
         }) {
             ic_cdk::spawn(join_group(self.group_id, args, self.attempt));